Add docusaurus dependency and a package.json script to create an initial, auto-generated docusaurus tree.
Add a script that edits the initially-generated docusaurus.config.js config
file.  Add acorn and acorn-walk dependencies used in this script.  This latter
part will be removed; it's not the right way to do this.  It seemed sensible
when it was a few edits, but now that the file is so different, it seems much
better to simply check in the desired default file.  This is a just-in-case
checkin, in case all or part of this approach is resurrected in future.